Add seL4_WriteBeg_fn and seL4_WriteEnd_fn.
These are used to any initialization/deinitialization of the seL4_Writer.
I use them to initialize the buffer index to 0 and append a 0 to the
the buffer when the writing is complete.

Changed radix 16 character to be lower case.

Fixed a bug where the leading zeros wasn't off by a factor of 2.